python/projects/video-subtitiles/backend/gcs_upload/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
google-cloud-storage==2.7.0
google-cloud-pubsub==2.13.0

python-multipart 

